+++ 8to10_frontend/src/api/PublicApi.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import CustomErrors from "@/api/CustomErrors.js";
-import {setAcceptHeaders, setContentTypeHeaders, setWithCredentials} from "@/helpers/AxiosConfig.js";
+import {setAcceptHeaders, setContentTypeHeaders, setWithCredentials} from "@/helpers/AxiosHelper.js";
 
 const publicApi = axios.create({
     baseURL: 'http://localhost:8080',
